wip: grades and updates of components;

+++ Assets/_Project/Scripts/Gameplay/Buildings/Building/Church/ChurchGradeData.cs
@@ -10,10 +10,10 @@ namespace _Project.Scripts.Gameplay.Buildings
         public int MaxFateStorageCapacity;
         public int MaxLightStorageCapacity = 1000;
         public int GradePrice;
-        public float TimeToProduceFate;
+        public int TimeToProduceFate;
         public int AmountToProduceFateAtTime;
         public float LightSendSpeed = 3f;
-        public float LightConsumeTime = 1f;
+        public int LightConsumeTime = 1;
         public int LightConsumeAmount = 1;
 
     }
